Moved mockData to separate file for legibility
